Moved tests to different class.
